+++ eslint.config.mjs
@@ -106,24 +106,16 @@ export default [
     files: [
       'scripts/createAndroidKeys.ts',
 
-      'scripts/loggingServer.ts',
-      'scripts/makeNativeHeaders.ts',
-      'scripts/patchFiles.ts',
-      'scripts/secretFiles.ts',
-
       'scripts/updateVersion.ts',
       'src/actions/BackupModalActions.tsx',
 
       'src/actions/CreateWalletActions.tsx',
-
-      'src/actions/DeviceSettingsActions.ts',
 
       'src/actions/FioActions.tsx',
       'src/actions/FioAddressActions.ts',
       'src/actions/FirstOpenActions.tsx',
       'src/actions/LoanWelcomeActions.tsx',
 
-      'src/actions/NotificationActions.ts',
       'src/actions/PaymentProtoActions.tsx',
       'src/actions/ReceiveDropdown.tsx',
       'src/actions/RecoveryReminderActions.tsx',
@@ -136,7 +128,6 @@ export default [
 
       'src/actions/WalletListActions.tsx',
 
-      'src/app.ts',
       'src/components/buttons/ButtonsView.tsx',
       'src/components/buttons/EdgeSwitch.tsx',
       'src/components/buttons/IconButton.tsx',
@@ -162,10 +153,9 @@ export default [
       'src/components/cards/TappableAccountCard.tsx',
       'src/components/cards/TappableCard.tsx',
       'src/components/cards/UnderlinedNumInputCard.tsx',
-      'src/components/cards/VisaCardCard.tsx',
+
       'src/components/cards/WalletRestoreCard.tsx',
       'src/components/cards/WarningCard.tsx',
-      'src/components/charts/SwipeChart.tsx',
 
       'src/components/common/AnimatedNumber.tsx',
       'src/components/common/BlurBackground.tsx',
@@ -306,7 +296,6 @@ export default [
       'src/components/scenes/PromotionSettingsScene.tsx',
 
       'src/components/scenes/SpendingLimitsScene.tsx',
-      'src/components/scenes/Staking/EarnScene.tsx',
 
       'src/components/scenes/SwapSettingsScene.tsx',
       'src/components/scenes/SwapSuccessScene.tsx',
@@ -329,11 +318,9 @@ export default [
       'src/components/services/NetworkActivity.ts',
       'src/components/services/PasswordReminderService.ts',
       'src/components/services/PermissionsManager.tsx',
-      'src/components/services/Providers.tsx',
 
       'src/components/services/SortedWalletList.ts',
       'src/components/services/StatusBarManager.tsx',
-      'src/components/services/WalletConnectService.tsx',
       'src/components/services/WalletLifecycle.ts',
       'src/components/services/WipeLogsService.tsx',
 
@@ -360,7 +347,7 @@ export default [
       'src/components/themed/LineTextDivider.tsx',
       'src/components/themed/MainButton.tsx',
       'src/components/themed/ManageTokensRow.tsx',
-      'src/components/themed/MenuTabs.tsx',
+
       'src/components/themed/ModalParts.tsx',
       'src/components/themed/PinDots.tsx',
 
@@ -403,11 +390,10 @@ export default [
       'src/components/tiles/PercentageChangeArrowTile.tsx',
       'src/components/tiles/TotalDebtCollateralTile.tsx',
 
-      'src/controllers/action-queue/ActionQueueStore.ts',
       'src/controllers/action-queue/cleaners.ts',
       'src/controllers/action-queue/push.ts',
       'src/controllers/action-queue/runtime/evaluateAction.ts',
-      'src/controllers/action-queue/runtime/executeActionProgram.ts',
+
       'src/controllers/edgeProvider/client/edgeProviderBridge.ts',
       'src/controllers/edgeProvider/client/pendingList.ts',
 
@@ -449,13 +435,8 @@ export default [
 
       'src/plugins/gui/providers/bityProvider.ts',
 
-      'src/plugins/gui/providers/mtpelerinProvider.ts',
-
       'src/plugins/gui/providers/revolutProvider.ts',
       'src/plugins/gui/RewardsCardPlugin.tsx',
-
-      'src/plugins/gui/util/fetchRevolut.ts',
-      'src/plugins/gui/util/initializeProviders.ts',
 
       'src/plugins/stake-plugins/generic/pluginInfo/optimismTarotPool.ts',
       'src/plugins/stake-plugins/generic/policyAdapters/CardanoKilnAdaptor.ts',
@@ -467,7 +448,6 @@ export default [
       'src/plugins/stake-plugins/generic/util/KilnApi.ts',
       'src/plugins/stake-plugins/generic/util/tarotUtils.ts',
       'src/plugins/stake-plugins/metadataCache.ts',
-      'src/plugins/stake-plugins/stakePlugins.ts',
 
       'src/plugins/stake-plugins/uniswapV2/Ecosystem.ts',
 
@@ -492,19 +472,17 @@ export default [
       'src/util/crypto.ts',
       'src/util/CryptoAmount.ts',
       'src/util/cryptoTextUtils.ts',
-      'src/util/CurrencyInfoHelpers.ts',
+
       'src/util/CurrencyWalletHelpers.ts',
 
       'src/util/exchangeRates.ts',
 
-      'src/util/FioAddressUtils.ts',
       'src/util/getAccountUsername.ts',
       'src/util/GuiPluginTools.ts',
       'src/util/haptic.ts',
       'src/util/infoUtils.ts',
 
       'src/util/memoUtils.ts',
-      'src/util/middleware/perfLogger.ts',
 
       'src/util/otpReminder.tsx',
       'src/util/scaling.ts',